+++ src/onegov.swissvotes/onegov/swissvotes/layouts/__init__.py
@@ -1,5 +1,7 @@
 from onegov.swissvotes.layouts.default import DefaultLayout
 from onegov.swissvotes.layouts.mail import MailLayout
+from onegov.swissvotes.layouts.page import AddPageLayout
+from onegov.swissvotes.layouts.page import DeletePageLayout
 from onegov.swissvotes.layouts.page import EditPageLayout
 from onegov.swissvotes.layouts.page import PageLayout
 from onegov.swissvotes.layouts.vote import DeleteVoteLayout
@@ -12,7 +14,9 @@ from onegov.swissvotes.layouts.votes import VotesLayout
 
 
 __all__ = (
+    'AddPageLayout',
     'DefaultLayout',
+    'DeletePageLayout',
     'DeleteVoteLayout',
     'DeleteVotesLayout',
     'EditPageLayout',
@@ -21,6 +25,6 @@ __all__ = (
     'UpdateVotesLayout',
     'UploadVoteAttachemtsLayout',
     'VoteLayout',
-    'VoteStrengthsLayout',
     'VotesLayout',
+    'VoteStrengthsLayout',
 )