+++ src/onegov.swissvotes/onegov/swissvotes/tests/conftest.py
@@ -20,7 +20,6 @@ def create_swissvotes_app(request, temporary_path):
         depot_backend='depot.io.local.LocalFileStorage',
         depot_storage_path=str(temporary_path),
     )
-    app.add_initial_content()
     app.session_manager.set_locale('de_CH', 'de_CH')
 
     session = app.session()