+++ src/onegov.swissvotes/onegov/swissvotes/views/principal.py
@@ -1,7 +1,6 @@
 from onegov.core.security import Public
 from onegov.swissvotes import SwissvotesApp
 from onegov.swissvotes.collections import TranslatablePageCollection
-from onegov.swissvotes.layouts import DefaultLayout
 from onegov.swissvotes.models import Principal
 
 
@@ -13,10 +12,5 @@ from onegov.swissvotes.models import Principal
 def view_home(self, request):
     """ The home page. """
 
-    page = TranslatablePageCollection(request.session).by_id('home')
-    if page:
-        return request.redirect(request.link(page))
-
-    return {
-        'layout': DefaultLayout(self, request),
-    }
+    page = TranslatablePageCollection(request.session).setdefault('home')
+    return request.redirect(request.link(page))